% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_functions.R
\name{f_germ_sum}
\alias{f_germ_sum}
\title{Calculate germination indices}
\usage{
f_germ_sum(SeedN = "NSeeds", evalName = "D", data)
}
\arguments{
\item{SeedN}{Number of seeds per replicate, or dataframe column with seed number per dish}

\item{evalName}{A character that prepend scoring time}

\item{data}{dataframe with data}
}
\description{
Calculate germination indices
}
