% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal_germinaR.R
\name{f_ngs}
\alias{f_ngs}
\title{GerminaR functions (Lozano-Isla et al., 2019) adapted and modified to work with decimal scoring times and allow for users to  choose the letters to prepend to scoring times.}
\usage{
f_ngs(evalName, data)
}
\arguments{
\item{evalName}{A character that prepend scoring time}

\item{data}{dataframe with data}
}
\description{
Number of germinated seeds
}
