% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_functions.R
\name{f_sdt}
\alias{f_sdt}
\title{Standard deviation of the Mean Germination Time}
\usage{
f_sdt(evalName, data)
}
\arguments{
\item{evalName}{A character that prepend scoring time}

\item{data}{dataframe with data}
}
\description{
Standard deviation of the Mean Germination Time
}
