% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_functions.R
\name{f_syn}
\alias{f_syn}
\title{Germination Synchronization Index, from Ranal, M. A., & Santana, D. G. D. (2006)}
\usage{
f_syn(evalName, data)
}
\arguments{
\item{evalName}{A character that prepend scoring time}

\item{data}{dataframe with data}
}
\description{
= 1 when germination of all the seeds occurs at the same time and = 0 when at least two seeds can germinate one each time. Thus, the value of synchrony assessments is the grade of overlap between seed germination.
}
