% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal_germinaR.R
\name{f_unc}
\alias{f_unc}
\title{Germination uncertainty, from Labouriau, G. (1983) adapted and modified to work with decimal scoring times and allow for users to  choose the letters to prepend to scoring times.}
\usage{
f_unc(evalName, data)
}
\arguments{
\item{evalName}{A character that prepend scoring time}

\item{data}{dataframe with data}
}
\description{
Adaptation of Shannon index measures the degree of uncertainty in predicting the informational entropy or uncertainty associated with the distribution of the relative frequency of germination (GOUVEA LABOURIAU 1983; LABOURIAU; VALADARES, 1983). Low values indicate frequencies with short peaks, i.e. the more concentrated the germination in time.
}
