% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rh_mix.R
\name{plot_rh_mix}
\alias{plot_rh_mix}
\title{Plots a basic graph showing LiCl after mixing two solutions}
\usage{
plot_rh_mix(rh1, vol1, rh2, vol2, theme = "light", units = "g/L")
}
\arguments{
\item{rh1}{The relative humidity percentage value of solution number one}

\item{vol1}{The volume 'in Litres' of solution rh1}

\item{rh2}{The relative humidity percentage value of solution number two}

\item{vol2}{The volume 'in Litres' of solution rh2#'}

\item{theme}{chose between light and dark plot background}

\item{units}{select LiCl concentration to be "g/L" or "g/100ml)}
}
\value{
a plot
}
\description{
Plots a basic graph showing LiCl after mixing two solutions
}
\examples{
plot_rh_mix(40,3,60,1)
}
