% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rh_scratch.R
\name{plot_rh_scratch}
\alias{plot_rh_scratch}
\title{Plots a basic graph showing desired relative humidity vs LiCl concentration}
\usage{
plot_rh_scratch(desired_rh, theme = "light")
}
\arguments{
\item{desired_rh}{The percentage value of the desired relative humidity}

\item{theme}{chose between light and dark plot background}
}
\value{
a plot
}
\description{
Plots a basic graph showing desired relative humidity vs LiCl concentration
}
\examples{
plot_rh_scratch(50,theme="dark")
}
