% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_functions.R
\name{rep_row}
\alias{rep_row}
\title{Variance of the Mean Germination Time}
\usage{
rep_row(Rseq, Nrow)
}
\arguments{
\item{Rseq}{no idea}

\item{Nrow}{no idea}
}
\description{
Variance of the Mean Germination Time
}
