% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_estimate.R
\name{rh_estimate}
\alias{rh_estimate}
\title{Given a weight in g. of LiCl and a volume, calculate relative humidity}
\usage{
rh_estimate(licl_g, volumen, verbose = TRUE)
}
\arguments{
\item{licl_g}{Ammount of LiCl g. in solution}

\item{volumen}{Solution volume 'in Litres'.}

\item{verbose}{Logical. If TRUE, suppresses messages or output during function execution.}
}
\value{
a data frame stating lithium chloride and units
}
\description{
Given a weight in g. of LiCl and a volume, calculate relative humidity
}
\examples{
rh_estimate(licl_g=300,volume=.75)
}
