% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_mix.R
\name{rh_mix}
\alias{rh_mix}
\title{Calculates the result relative humidity of mixing two lithium chloride solutions}
\usage{
rh_mix(rh1, vol1, rh2, vol2)
}
\arguments{
\item{rh1}{The relative humidity percentage value of solution number one}

\item{vol1}{The volume 'in Litres' of solution rh1}

\item{rh2}{The relative humidity percentage value of solution number two}

\item{vol2}{The volume 'in Litres' of solution rh2}
}
\value{
a string with mixed solution relative humidity percentage
}
\description{
Calculates the result relative humidity of mixing two lithium chloride solutions
}
\examples{
rh_mix(rh1=50,vol1=0.5,rh2=70,vol2=0.25)

}
