% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_rh.R
\name{rh_modify}
\alias{rh_modify}
\title{Calculate neccesary changes to lithium chloride solutions to achieve target relative humidity}
\usage{
rh_modify(initial_rh, desired_rh, initial_volume)
}
\arguments{
\item{initial_rh}{The relative humidity percentage value of the solution to be modified}

\item{desired_rh}{The percentage value of the desired relative humidity}

\item{initial_volume}{The volume 'in Litres' of the initial solution}
}
\value{
a string with instructions to modify the solution
}
\description{
Obtain T50 values using Coolbear et al. (1984) formula modified by Farooq et al.(2005). Vector containing scoring days must have same length than vector containing cumulative germination data
}
\examples{
rh_modify(initial_rh=20,desired_rh=40,initial_volume=0.5)
}
