% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_scratch.R
\name{rh_scratch}
\alias{rh_scratch}
\title{Given a relative humidity and a volume of water, returns the lithium chloride necessary to reach it}
\usage{
rh_scratch(desired_rh, volume, verbose = TRUE)
}
\arguments{
\item{desired_rh}{The percentage value of the desired relative humidity}

\item{volume}{The desired volume 'in Litres'.}

\item{verbose}{Logical. If TRUE, suppresses messages or output during function execution.}
}
\value{
a data frame stating lithium chloride and units
}
\description{
Given a relative humidity and a volume of water, returns the lithium chloride necessary to reach it
}
\examples{
rh_scratch(desired_rh=60,volume=0.5)
}
