% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgp_corner_temp.R
\name{tgp_corner_temp}
\alias{tgp_corner_temp}
\title{Finds mean of stables temperatures for probes}
\usage{
tgp_corner_temp(file)
}
\arguments{
\item{file}{A character string specifying the path to the CSV file.}
}
\value{
A data frame
}
\description{
Reads and inds mean of stables temperatures for probes
}
