% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgp_squirrel.R
\name{tgp_squirrel}
\alias{tgp_squirrel}
\title{Transform Squirrel Temperature Data}
\usage{
tgp_squirrel(file)
}
\arguments{
\item{file}{A character string specifying the path to the CSV file.}
}
\value{
A data frame with columns for the original data and two new columns:
\code{Date}Data collection timestamp, \code{Probe} indicating the probe identifier and \code{Temperature} containing the corresponding readings.
}
\description{
Reads a CSV file with temperature gradient data from a Squirrel logger, skips the first 17 lines, and reshapes the data from wide to long format,
extracting temperature readings from multiple probes.
}
