% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgp_squirrel_plot.R
\name{tgp_squirrel_plot}
\alias{tgp_squirrel_plot}
\title{Plot Squirrel Temperature Data}
\usage{
tgp_squirrel_plot(file, start = NULL, end = NULL)
}
\arguments{
\item{file}{A character string specifying the path to the CSV file.}
}
\value{
A plot
}
\description{
Reads a CSV file and plots temperature gradient plate data
}
