% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_modify.R
\name{rh_modify}
\alias{rh_modify}
\title{Calculate necessary changes to lithium chloride solutions to achieve target relative humidity}
\usage{
rh_modify(initial_rh, desired_rh, initial_volume, verbose = TRUE)
}
\arguments{
\item{initial_rh}{The relative humidity percentage value of the solution to be modified}

\item{desired_rh}{The percentage value of the desired relative humidity}

\item{initial_volume}{The volume 'in Litres' of the initial solution}

\item{verbose}{Logical. If TRUE, suppresses messages or output during function execution.}
}
\value{
a string with instructions to modify the solution
}
\description{
Calculate necessary changes to lithium chloride solutions to achieve target relative humidity
}
\examples{
rh_modify(initial_rh=20,desired_rh=40,initial_volume=0.5)
}
