% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyTag_plot.R
\name{TinyTag_plot}
\alias{TinyTag_plot}
\title{TinyTag Plot}
\usage{
TinyTag_plot(file_path, title = "TinyTag Data")
}
\arguments{
\item{title}{Plot title. Defaults to "TinyTag Data".}

\item{table}{Object with Path to the CSV file to read.}
}
\value{
A ggplot object.
}
\description{
This function reads TinyTag CSV data (e.g. from growth chambers),
cleans it, and generates a dual-axis ggplot showing temperature and humidity.
}
\examples{
\dontrun{
TinyTag_plot("Growth chamber/Large_growth_chamber.csv", title = "Large Growth Chamber")
}
}
